Bucket commit

+++ Chowdren/base/objects/easing.h
@@ -4,7 +4,7 @@
 class EasingObject
 {
 public:
-    static double ease_out(int a, int b, int function, double step);
+    static float ease_out(int a, int b, int function, double step);
 };
 
 #endif // CHOWDREN_EASING_H
